requirements.txt
torch
numpy
sympy
aiohttp
ray
matplotlib
pytest
python-dotenv
requests
urllib3
pandas